relocate client and logger
