+++ core/src/service/api_service.rs
@@ -91,7 +91,7 @@ mod request_common_tests {
 
     use crate::assert_matches;
     use crate::model::state::temp_config;
-    use crate::service::client::{request_helper, ApiError};
+    use crate::service::api_service::{request_helper, ApiError};
     use crate::service::db_state_service::get_code_version;
     use crate::service::test_utils;
