+++ core/src/service/log_service.rs
