+++ core/src/service/mod.rs
@@ -1,5 +1,5 @@
 pub mod account_service;
-pub mod client;
+pub mod api_service;
 pub mod db_state_service;
 pub mod drawing_service;
 pub mod file_compression_service;
@@ -7,6 +7,7 @@ pub mod file_encryption_service;
 pub mod file_service;
 pub mod import_export_service;
 pub mod integrity_service;
+pub mod log_service;
 pub mod path_service;
 pub mod sync_service;
 pub mod test_utils;
