+++ core/tests/create_account_tests.rs
@@ -2,7 +2,7 @@ mod integration_test;
 
 #[cfg(test)]
 mod change_document_content_tests {
-    use lockbook_core::service::client;
+    use lockbook_core::service::api_service;
     use lockbook_core::service::test_utils::{generate_account, generate_root_metadata};
     use lockbook_models::api::*;
 
@@ -11,6 +11,6 @@ mod change_document_content_tests {
         // new account
         let account = generate_account();
         let (root, _root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
     }
 }
