+++ core/tests/delete_file_tests.rs
@@ -4,8 +4,8 @@ mod integration_test;
 mod delete_document_tests {
     use lockbook_core::assert_get_updates_required;
     use lockbook_core::assert_matches;
-    use lockbook_core::service::client;
-    use lockbook_core::service::client::ApiError;
+    use lockbook_core::service::api_service;
+    use lockbook_core::service::api_service::ApiError;
     use lockbook_core::service::test_utils::{
         generate_account, generate_file_metadata, generate_root_metadata,
     };
@@ -18,12 +18,12 @@ mod delete_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create document
         let (mut doc, _doc_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Document);
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new(&doc)],
@@ -33,7 +33,7 @@ mod delete_document_tests {
 
         // delete document
         doc.deleted = true;
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new_diff(root.id, &doc.name, &doc)],
@@ -47,12 +47,12 @@ mod delete_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create document
         let (doc, _doc_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Document);
-        let result = client::request(
+        let result = api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 // create document as if deleting an existing document
@@ -67,12 +67,12 @@ mod delete_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create document
         let (mut doc, _doc_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Document);
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new(&doc)],
@@ -82,7 +82,7 @@ mod delete_document_tests {
 
         // delete document
         doc.deleted = true;
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new_diff(root.id, &doc.name, &doc)],
@@ -91,7 +91,7 @@ mod delete_document_tests {
         .unwrap();
 
         // delete document again
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new_diff(root.id, &doc.name, &doc)],
@@ -105,11 +105,11 @@ mod delete_document_tests {
         // new account
         let account = generate_account();
         let (mut root, _root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // delete root
         root.deleted = true;
-        let result = client::request(
+        let result = api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 // create document as if deleting an existing document
