+++ core/tests/get_updates_test.rs
@@ -2,7 +2,7 @@ mod integration_test;
 
 #[cfg(test)]
 mod get_updates_test {
-    use lockbook_core::service::client;
+    use lockbook_core::service::api_service;
     use lockbook_core::service::test_utils::{generate_account, generate_root_metadata};
     use lockbook_models::api::{GetUpdatesRequest, NewAccountRequest};
 
@@ -11,12 +11,13 @@ mod get_updates_test {
         // new account
         let account = generate_account();
         let (mut root, _) = generate_root_metadata(&account);
-        root.metadata_version = client::request(&account, NewAccountRequest::new(&account, &root))
-            .unwrap()
-            .folder_metadata_version;
+        root.metadata_version =
+            api_service::request(&account, NewAccountRequest::new(&account, &root))
+                .unwrap()
+                .folder_metadata_version;
 
         // get updates at version 0
-        let result = client::request(
+        let result = api_service::request(
             &account,
             GetUpdatesRequest {
                 since_metadata_version: 0,
@@ -28,7 +29,7 @@ mod get_updates_test {
         assert_eq!(result, 1);
 
         // get updates at version of root folder
-        let result = client::request(
+        let result = api_service::request(
             &account,
             GetUpdatesRequest {
                 since_metadata_version: root.metadata_version,
