+++ core/tests/move_file_tests.rs
@@ -2,8 +2,8 @@ mod integration_test;
 
 #[cfg(test)]
 mod move_document_tests {
-    use lockbook_core::service::client;
-    use lockbook_core::service::client::ApiError;
+    use lockbook_core::service::api_service;
+    use lockbook_core::service::api_service::ApiError;
     use lockbook_core::service::test_utils::{
         generate_account, generate_file_metadata, generate_root_metadata,
     };
@@ -17,14 +17,14 @@ mod move_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create document and folder
         let (mut doc, _doc_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Document);
         let (folder, _folder_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Folder);
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new(&doc), FileMetadataDiff::new(&folder)],
@@ -34,7 +34,7 @@ mod move_document_tests {
 
         // move document
         doc.parent = folder.id;
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new_diff(root.id, &doc.name, &doc)],
@@ -48,14 +48,14 @@ mod move_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create document and folder
         let (folder, _folder_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Folder);
         let (doc, _doc_key) =
             generate_file_metadata(&account, &folder, &root_key, FileType::Document);
-        let result = client::request(
+        let result = api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 // create document as if moving an existing document
@@ -73,14 +73,14 @@ mod move_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create document and folder, but don't send folder to server
         let (folder, _folder_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Folder);
         let (doc, _doc_key) =
             generate_file_metadata(&account, &folder, &root_key, FileType::Document);
-        let result = client::request(
+        let result = api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new(&doc)],
@@ -94,7 +94,7 @@ mod move_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create deleted document and folder
         let (mut doc, _doc_key) =
@@ -102,7 +102,7 @@ mod move_document_tests {
         doc.deleted = true;
         let (folder, _folder_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Folder);
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new(&doc), FileMetadataDiff::new(&folder)],
@@ -112,7 +112,7 @@ mod move_document_tests {
 
         // move document
         doc.parent = folder.id;
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new_diff(root.id, &doc.name, &doc)],
@@ -126,14 +126,14 @@ mod move_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create document and folder
         let (mut doc, _doc_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Document);
         let (folder, _folder_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Folder);
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new(&doc), FileMetadataDiff::new(&folder)],
@@ -143,7 +143,7 @@ mod move_document_tests {
 
         // move document
         doc.parent = folder.id;
-        let result = client::request(
+        let result = api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 // use incorrect previous parent
@@ -158,7 +158,7 @@ mod move_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create documents and folder
         let (folder, _folder_key) =
@@ -168,7 +168,7 @@ mod move_document_tests {
         let (mut doc2, _doc_key2) =
             generate_file_metadata(&account, &folder, &root_key, FileType::Document);
         doc2.name = doc.name.clone();
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![
@@ -182,7 +182,7 @@ mod move_document_tests {
 
         // move document
         doc.parent = folder.id;
-        let result = client::request(
+        let result = api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new_diff(root.id, &doc.name, &doc)],
@@ -196,12 +196,12 @@ mod move_document_tests {
         // new account
         let account = generate_account();
         let (mut root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create folder
         let (folder, _folder_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Folder);
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new(&folder)],
@@ -211,7 +211,7 @@ mod move_document_tests {
 
         // move root
         root.parent = folder.id;
-        let result = client::request(
+        let result = api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new_diff(root.id, &root.name, &root)],
@@ -225,12 +225,12 @@ mod move_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create folder
         let (mut folder, _folder_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Folder);
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new(&folder)],
@@ -240,7 +240,7 @@ mod move_document_tests {
 
         // move folder into self
         folder.parent = folder.id;
-        let result = client::request(
+        let result = api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new_diff(root.id, &folder.name, &folder)],
@@ -254,14 +254,14 @@ mod move_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create folders
         let (mut folder, _folder_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Folder);
         let (folder2, _folder_key2) =
             generate_file_metadata(&account, &folder, &root_key, FileType::Folder);
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![
@@ -274,7 +274,7 @@ mod move_document_tests {
 
         // move folder into itself
         folder.parent = folder2.id;
-        let result = client::request(
+        let result = api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new_diff(root.id, &folder.name, &folder)],
@@ -288,14 +288,14 @@ mod move_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create documents
         let (mut doc, _doc_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Document);
         let (doc2, _doc_key2) =
             generate_file_metadata(&account, &root, &root_key, FileType::Document);
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new(&doc), FileMetadataDiff::new(&doc2)],
@@ -305,7 +305,7 @@ mod move_document_tests {
 
         // move folder into itself
         doc.parent = doc2.id;
-        let result = client::request(
+        let result = api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new_diff(root.id, &doc.name, &doc)],
