+++ core/tests/new_account_tests.rs
@@ -3,8 +3,8 @@ mod integration_test;
 #[cfg(test)]
 mod new_account_tests {
     use lockbook_core::assert_matches;
-    use lockbook_core::service::client;
-    use lockbook_core::service::client::ApiError;
+    use lockbook_core::service::api_service;
+    use lockbook_core::service::api_service::ApiError;
     use lockbook_core::service::test_utils::{generate_account, generate_root_metadata};
     use lockbook_models::api::*;
 
@@ -12,16 +12,16 @@ mod new_account_tests {
     fn new_account() {
         let account = generate_account();
         let (root, _) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
     }
 
     #[test]
     fn new_account_duplicate() {
         let account = generate_account();
         let (root, _) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
-        let result = client::request(&account, NewAccountRequest::new(&account, &root));
+        let result = api_service::request(&account, NewAccountRequest::new(&account, &root));
         assert_matches!(
             result,
             Err(ApiError::<NewAccountError>::Endpoint(
@@ -34,12 +34,12 @@ mod new_account_tests {
     fn new_account_duplicate_username() {
         let account = generate_account();
         let (root, _) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         let mut account2 = generate_account();
         account2.username = account.username;
         let (root2, _) = generate_root_metadata(&account2);
-        let result = client::request(&account2, NewAccountRequest::new(&account2, &root2));
+        let result = api_service::request(&account2, NewAccountRequest::new(&account2, &root2));
         assert_matches!(
             result,
             Err(ApiError::<NewAccountError>::Endpoint(
@@ -54,7 +54,7 @@ mod new_account_tests {
         account.username += " ";
         let (root, _) = generate_root_metadata(&account);
 
-        let result = client::request(&account, NewAccountRequest::new(&account, &root));
+        let result = api_service::request(&account, NewAccountRequest::new(&account, &root));
         assert_matches!(
             result,
             Err(ApiError::<NewAccountError>::Endpoint(
