+++ core/tests/rename_file_tests.rs
@@ -4,8 +4,8 @@ mod integration_test;
 mod rename_document_tests {
     use lockbook_core::assert_get_updates_required;
     use lockbook_core::assert_matches;
-    use lockbook_core::service::client;
-    use lockbook_core::service::client::ApiError;
+    use lockbook_core::service::api_service;
+    use lockbook_core::service::api_service::ApiError;
     use lockbook_core::service::test_utils::{
         generate_account, generate_file_metadata, generate_root_metadata, random_filename,
     };
@@ -18,12 +18,12 @@ mod rename_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create document
         let (mut doc, _doc_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Document);
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new(&doc)],
@@ -34,7 +34,7 @@ mod rename_document_tests {
         // rename document
         let old_name = doc.name.clone();
         doc.name = random_filename();
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new_diff(root.id, &old_name, &doc)],
@@ -48,12 +48,12 @@ mod rename_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create document
         let (doc, _doc_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Document);
-        let result = client::request(
+        let result = api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 // create document as if renaming an existing document
@@ -68,13 +68,13 @@ mod rename_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create document
         let (mut doc, _doc_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Document);
         doc.deleted = true;
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new(&doc)],
@@ -85,7 +85,7 @@ mod rename_document_tests {
         // rename document
         let old_name = doc.name.clone();
         doc.name = random_filename();
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new_diff(root.id, &old_name, &doc)],
@@ -99,12 +99,12 @@ mod rename_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create document
         let (mut doc, _doc_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Document);
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new(&doc)],
@@ -114,7 +114,7 @@ mod rename_document_tests {
 
         // rename document
         doc.name = random_filename();
-        let result = client::request(
+        let result = api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 // use incorrect previous name
@@ -129,14 +129,14 @@ mod rename_document_tests {
         // new account
         let account = generate_account();
         let (root, root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // create 2 document
         let (mut doc, _doc_key) =
             generate_file_metadata(&account, &root, &root_key, FileType::Document);
         let (doc2, _doc_key2) =
             generate_file_metadata(&account, &root, &root_key, FileType::Document);
-        client::request(
+        api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new(&doc), FileMetadataDiff::new(&doc2)],
@@ -147,7 +147,7 @@ mod rename_document_tests {
         // rename document to match name of other document
         let old_name = doc.name.clone();
         doc.name = doc2.name;
-        let result = client::request(
+        let result = api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new_diff(root.id, &old_name, &doc)],
@@ -161,10 +161,10 @@ mod rename_document_tests {
         // new account
         let account = generate_account();
         let (root, _root_key) = generate_root_metadata(&account);
-        client::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
+        api_service::request(&account, NewAccountRequest::new(&account, &root)).unwrap();
 
         // rename root
-        let result = client::request(
+        let result = api_service::request(
             &account,
             FileMetadataUpsertsRequest {
                 updates: vec![FileMetadataDiff::new_diff(root.id, &root.name, &root)],
